requirements.txt
autopep8==1.5.7
boto3==1.18.64
botocore==1.21.64
channels==3.0.4
channels-redis==3.3.1
daphne==3.0.2
Django==3.2.8
django-environ==0.8.0
django-storages==1.12.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
mysqlclient==2.0.3
Pillow==8.4.0
pymongo==3.12.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7

